+++ typed-builder-macro/src/lib.rs
@@ -30,9 +30,10 @@ fn impl_my_derive(ast: &syn::DeriveInput) -> Result<TokenStream, Error> {
                     .filter(|f| f.builder_attr.default.is_none())
                     .map(|f| struct_info.required_field_impl(f));
                 let mutators = struct_info
-                    .builder_attr
-                    .mutators
+                    .fields
                     .iter()
+                    .flat_map(|f| &f.builder_attr.mutators)
+                    .chain(&struct_info.builder_attr.mutators)
                     .map(|m| struct_info.mutator_impl(m))
                     .collect::<Result<TokenStream, _>>()?;
                 let build_method = struct_info.build_method_impl();